fix failing unit tests and move new unit tests into existing file i had not noticed
